chore: refactor folder structure for dao domains + consts
