+++ src/daos/cryptex/tokenCDTest/constants/tokenCDTest.ts
@@ -1,0 +1,8 @@
+// TODO: Remove tokenCDTest when mainnet capital distributor is live (APP-558)
+export const tokenCDTest = {
+    id: 'ethereum-sepolia-0x3c5909c6671E012950aE5fFe1578ffe828A53711',
+    name: 'tokenCDTest',
+};
+
+export const tokenCDTestTokenVotingPluginAddress: `0x${string}` =
+    '0x1652FDd272fEf49B53bd102550DE775519e60b8E';
